requirements.txt
requests==2.32.3
google-generativeai==0.3.2
python-dotenv==1.0.0
pillow==9.5.0
flask==3.0.0 